make precision UInt8

+++ Sources/EOSIO/API/V2/Hyperion.swift
@@ -55,7 +55,7 @@ public extension API.V2.Hyperion {
 
         public struct Token: Decodable {
             public let symbol: String
-            public let precision: UInt
+            public let precision: UInt8
             public let amount: Double
             public let contract: Name
         }
